requirements.txt
customtkinter>=5.2.0
pywin32>=306
pycryptodome>=3.19.0
pyinstaller>=6.3.0
pillow>=10.0.0
requests>=2.31.0
